Initial commit
VSCode version of the MERGED_SCRIPT_WITH_ERROR_HANDLING.

The only input required in the script is the PDF file name in the "pdf_file_path" variable, on line 72.

Ensure that the PDF file is in the same folder as the script.

The rest of the script uses this variable to create the JSON files that are used as inputs by other functions.

Bugs : It doesn't take screenshots of text that is on a single line.